handle bespoke SyC error (i.e. not marshalled using grpc-status)
