+++ src/ansys/systemcoupling/core/client/services/solution.py
@@ -20,11 +20,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import ansys.api.systemcoupling.v0.error_pb2 as syc_error_pb2
 import ansys.api.systemcoupling.v0.solution_pb2 as solution_pb2
 import ansys.api.systemcoupling.v0.solution_pb2_grpc as solution_pb2_grpc
 import grpc
-from grpc_status.rpc_status import from_call
+
+from ansys.systemcoupling.core.client.services.handle_rpc_error import handle_rpc_error
 
 
 class SolutionService:
@@ -36,16 +36,7 @@ class SolutionService:
         try:
             self.__stub.Solve(request)
         except grpc.RpcError as rpc_error:
-            status = from_call(rpc_error)
-            msg = f"Command execution failed: {status.message} (code={status.code})"
-            for detail in status.details:
-                if detail.Is(syc_error_pb2.ErrorDetails.DESCRIPTOR):
-                    info = syc_error_pb2.ErrorDetails()
-                    detail.Unpack(info)
-                    msg += (
-                        f"\n\nServer exception details:\n"
-                        f"{info.exception_classname}\n{info.stack_trace}"
-                    )
+            msg = handle_rpc_error(rpc_error)
             raise RuntimeError(msg) from None
 
     def interrupt(self, reason):
